Don't put the build directory name into man page title
Also remove "Hoola" from the man page title and replace it with a
proper man page name.

Without this patch the the head line of the wml(1) man page (as
generated from POD by cmake/pod2man-wrapper.pl) looks like this:

  HOOLA(1)                     EN Tools                     HOOLA(1)

And the man page wml::des::imgbg(3) looks even worse:

  tmp::ho1p32hifo::Hoola(3)    EN Tools    tmp::ho1p32hifo::Hoola(3)

With this patch the man page titles look like they should:

  WML(1)                       EN Tools                       WML(1)

  wml::des::imgbg(3)           EN Tools           wml::des::imgbg(3)

Not having the build directory path in any of the resulting files is
additionally necessary to make the build reproducible. See
https://reproducible-builds.org/